Allow poll creators to use /removeoption
This also fixes a bug where only server admins could use !removeoption
but any user could use /removeoption for the latest poll.

This change means that removing the last vote from an option no longer
causes the bot to remove that option. Instead, /removeoption is now the
only way to remove an option.